u32 -> u64, adjust tolerance

+++ dkg-gadget/src/async_protocols/mod.rs
@@ -75,7 +75,7 @@ pub struct AsyncProtocolParameters<
 	pub best_authorities: Arc<Vec<(KeygenPartyId, Public)>>,
 	pub authority_public_key: Arc<Public>,
 	pub party_i: KeygenPartyId,
-	pub associated_block_id: u32,
+	pub associated_block_id: u64,
 	pub batch_id_gen: Arc<AtomicU64>,
 	pub handle: AsyncProtocolRemote<BI::Clock>,
 	pub session_id: SessionId,
@@ -290,27 +290,27 @@ pub enum ProtocolType<MaxProposalLength: Get<u32> + Clone + Send + Sync + std::f
 		i: KeygenPartyId,
 		t: u16,
 		n: u16,
-		associated_block_id: u32,
+		associated_block_id: u64,
 	},
 	Offline {
 		unsigned_proposal: Arc<UnsignedProposal<MaxProposalLength>>,
 		i: OfflinePartyId,
 		s_l: Vec<KeygenPartyId>,
 		local_key: Arc<LocalKey<Secp256k1>>,
-		associated_block_id: u32,
+		associated_block_id: u64,
 	},
 	Voting {
 		offline_stage: Arc<CompletedOfflineStage>,
 		unsigned_proposal: Arc<UnsignedProposal<MaxProposalLength>>,
 		i: OfflinePartyId,
-		associated_block_id: u32,
+		associated_block_id: u64,
 	},
 }
 
 impl<MaxProposalLength: Get<u32> + Clone + Send + Sync + std::fmt::Debug + 'static>
 	ProtocolType<MaxProposalLength>
 {
-	pub const fn get_associated_block_id(&self) -> u32 {
+	pub const fn get_associated_block_id(&self) -> u64 {
 		match self {
 			Self::Keygen { associated_block_id, .. } |
 			Self::Offline { associated_block_id, .. } |
